Always check jobs when loading jobs dir.
